requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
